\name{HTA_Location}
\alias{HTA_Location}
\docType{data}
\title{The coordinates of probes and the mapped PM probes for hta2.0 chips}
\description{
  This data include the probes location for hta2.0 chips.
}
\usage{data(HTA_Location)}
\format{
  A 1*5118823 matrix including the location for unique probes in HTA_transcript_NO.
}
\source{
Danielle Thierry-Mieg ,Jean Thierry-Mieg. Aceview: a comprehensive cDNA-supported gene and transcripts annotaion.
Genome Biology.2006,7(Suppl 1):S12

Manhong Dai, Pinglang Wang,Andrew D. Boyd. (2005) Evolving gene/transcript definitions significantly
alter the interpretation of GeneChip data,Nucleic Acid Research 33(20):e175.
}

\keyword{datasets}
