\name{HTA_probes_transcripts}
\alias{HTA_probes_transcripts}
\docType{data}
\title{ The number of probes and the number of transcripts mapped to each gene for hta2.0 chips}
\description{
  This data is the number of probes and the number of transcripts mapped to each gene for hta2.0 chips.
}
\usage{data(HTA_probes_transcripts)}
\format{
  A 33394*2 matrix including the number of probes and the number of transcripts mapped to each of 33394 genes for hta20 chips. 
 
}
\source{
Danielle Thierry-Mieg ,Jean Thierry-Mieg. Aceview: a comprehensive cDNA-supported gene and transcripts annotaion
Genome Biology.2006,7(Suppl 1):S12

Manhong Dai, Pinglang Wang,Andrew D. Boyd. (2005) Evolving gene/transcript definitions significantly
alter the interpretation of GeneChip data,Nucleic Acid Research 33(20):e175.
}

\keyword{datasets}
