\name{HTA_transcript_name}
\alias{HTA_transcript_name}
\docType{data}
\title{ The names of transcripts mapped to each gene for hta2.0 chips}
\description{
  This data include the names of transcripts mapped to each gene for hta2.0 chips.
}
\usage{data(HTA_transcript_name)}
\format{
  A 225456*1 matrix including 225456 transcript names mapped to genes for hta2.0 chips. 
 
}
\source{
Danielle Thierry-Mieg ,Jean Thierry-Mieg. Aceview: a comprehensive cDNA-supported gene and transcripts annotaion
Genome Biology.2006,7(Suppl 1):S12

Manhong Dai, Pinglang Wang,Andrew D. Boyd. (2005) Evolving gene/transcript definitions significantly
alter the interpretation of GeneChip data,Nucleic Acid Research 33(20):e175.
}

\keyword{datasets}
