\name{HTA_transcript_NO}
\alias{HTA_transcript_NO}
\docType{data}
\title{The coordinates of probes and the mapped transcripts for hta2.0 chips}
\description{
  This data include the coordinates of probes and the mapped transcripts for hta2.0 chips.
}
\usage{data(HTA_transcript_NO)}
\format{
  A 20626078*3 matrix including pos_x,pos_y and transcript_no.
  pos_x and pos_y are respectively X and Y coordinates of probes for hta2.0 chips.
  Transcript_no is the mapped transcripts for each probe.
 
}
\source{
Danielle Thierry-Mieg ,Jean Thierry-Mieg. Aceview: a comprehensive cDNA-supported gene and transcripts annotaion
Genome Biology.2006,7(Suppl 1):S12

Manhong Dai, Pinglang Wang,Andrew D. Boyd. (2005) Evolving gene/transcript definitions significantly
alter the interpretation of GeneChip data,Nucleic Acid Research 33(20):e175.
}

\keyword{datasets}
