\name{Human_Location}
\alias{Human_Location}
\docType{data}
\title{The coordinates of probes and the mapped PM for human exon chips}
\description{
  This data include the probes location for human exon chips.
}
\usage{data(Human_Location)}
\format{
   A 1*1565476 matrix including the location for unique probes in Human_transcript_NO.
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
