\name{Human_probes_transcripts}
\alias{Human_probes_transcripts}
\docType{data}
\title{ The number of probes and the number of transcripts mapped to each gene for human exon chips}
\description{
  This data is the number of probes and the number of transcripts mapped to each gene for human exon chips.
}
\usage{data(Human_probes_transcripts)}
\format{
  A 40174*2 matrix including the number of probes and the number of transcripts mapped to each of 40174 genes for human exon chips. 
 
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
