\name{Human_transcript_name}
\alias{Human_transcript_name}
\docType{data}
\title{ The names of transcripts mapped to each gene for human exon chips}
\description{
  This data include the names of transcripts mapped to each gene for human exon chips.
}
\usage{data(Human_transcript_name)}
\format{
  A 121741*1 matrix including 121741 transcript names mapped to genes for human exon chips. 
 
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
