\name{Mouse_probes_transcripts}
\alias{Mouse_probes_transcripts}
\docType{data}
\title{ The number of probes and the number of transcripts mapped to each gene for mouse exon chips }
\description{
   This data include the number of probes and the number of transcripts mapped to each gene for mouse exon chips.
}
\usage{data(Mouse_probes_transcripts)}
\format{
  A 27719*2 matrix including the number of probes and the number of transcripts mapped to each of 27719 genes for mouse exon chips.
 
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
