\name{Mouse_transcript_name}
\alias{Mouse_transcript_name}
\docType{data}
\title{ The names of transcripts mapped to each gene for mouse exon chips}
\description{
   This data include the names of transcripts mapped to each gene for mouse exon chips
}
\usage{data(Mouse_transcript_name)}
\format{
  A 75751*1 matrix including 75751 transcript names mapped to genes for mouse exon chips. 
 
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
