\name{Mouse_transcript_NO}
\alias{Mouse_transcript_NO}
\docType{data}
\title{ The coordinates of probes and the mapped transcripts mapped for mouse exon chips}
\description{
 This data include the coordinates of probes and the mapped transcripts for mouse exon chips.
}
\usage{data(Mouse_transcript_NO)}
\format{
  A 2928848*3 matrix including pos_x,pos_y and transcript_no. 
  pos_x and pos_y are respectively X and Y coordinates of probes for mouse exon chips.
  Transcript_no data is the mapped transcripts for each probe.
 
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
