\name{Rat_Location}
\alias{Rat_Location}
\docType{data}
\title{The coordinates of probes and the mapped PM for Rat exon chips}
\description{
  This data include the probes location for Rat exon chips.
}
\usage{data(Rat_Location)}
\format{
   A 1*931210 matrix including the location for unique probes in Rat_transcript_NO.
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
