\name{Rat_transcript_name}
\alias{Rat_transcript_name}
\docType{data}
\title{ The names of transcripts mapped to each gene for rat exon chips}
\description{
   This data is the names of transcripts mapped to each gene for rat exon chips
}
\usage{data(Rat_transcript_name)}
\format{
  A 334851*1 matrix including 334851 transcript names mapped to each gene for rat exon chips. 
 
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
