\name{Rat_transcript_NO}
\alias{Rat_transcript_NO}
\docType{data}
\title{ The coordinates of probes and the mapped transcripts for rat exon chips}
\description{
  This data include the coordinates of probes and the mapped transcripts for rat exon chips.
}
\usage{data(Rat_transcript_NO)}
\format{
  A 1491570*3 matrix including pos_x,pos_y and transcript_no.
  pos_x and pos_y are respectively X and Y coordinates of probes for rat exon chips.
  Transcript_no is the mapped transcripts for each probe. 
 
}
\source{
Risueno A, Fontanillo C, Dinger ME, De Las Rivas J. GATExplorer: genomic and transcriptomic explorer; mapping expression
probes to gene loci, transcripts, exons and ncRNAs. BMC Bioinformatics.2010 Apr 29;11:221.

}

\keyword{datasets}
